day 2 finished and extras
Finished day 2! Had an assertion error but I fixed that by resetting the vector before I break from the nested for loop in findNounAndVerb(). Cleaned up the code so anyone looking at it can try it out for themselves and see how things work. I used quite a bit of functions. Not sure if this is good habits or not as I'm not too sure if I used pass by reference currectly. Either way, the code works and runs without error. NOTE: Both  solutions can't be ran at the same time. Not a big deal and I feel as if i can solve that issue but it should never happen to begin with soooo OOF. My brain is mush. This is my final answer.