got the range working
was able to convert niko's solution to vectors so far. I managed to display all values between points and save them into a vector that stores points. Now I need to just compare the vector of points of the black and red wires. That shouldn't be too difficult but my brain is done with this thing for now. It was a headache and a half to get this far. oof