almost there
I am almost there. I am able to get an answer so far. The logic that I written seems sound enough to me. I need to review my new way of finding the extra bit of segment to add or remove. Also need to figure out how to get the stupid 0,0 coordinate out of the points on interest. That shit is annoying.