day 3 part 2 finished
I finished part 2 of day 3 finally. This is exactly how the code looked when I completed it. "Source.cpp" is a solution found online that I forget the author. It kind of inspired me, or at least made me lean toward the right direction, sort of. It mostly served as a reminder that my code could be significantly smaller and that my previous solution was probably too complex. I am super glad I started from a fresh version of my first part solution. All I was missing was the distance formula. I wasn't properly getting the distance between two points to subtract. I remembered that the distance formula is a thing so there you go. The literal last piece of the puzzle was something I learned in grade school. BIG OOOOOOF