Cleaning up
Just cleaning up things. This has the old day3.cpp that had my inital solution to day3 part 1. I left it in here becase it still has the explanations in it. Maybe somebody will use curious enough to dig through and find the explanation. The next commit will be the new day3.cpp with the part 2 solution in it. I deleted the explanation because I want the code to look less intimidating and take up less lines.