day 4 part 1 finished
Finished this part of the day a lot faster than I thought I would. A nice change of pace from day 3 which took nearly a month to solve. This part 1 will more than likely be altered to have less code by converting my adj and inc functions to bool and having them used when looping through the range of numbers.