Day 4 Part 2 Finished
Finished Day 4 part 2. This is how the code looke exactly when I got the solution...sorta. I am pushing this version so I can keep it in case I mess up. I am going to convert the functions I have to bool and see if that makes this thing a litte faster since I won't be looping through so much crap. Fingers crossed.