Finished
Day 4 parts 1 and 2 are done. I managed to translate the code to work with my bool idea and significantly increased the speed, at least it looked like it increases. Feels like it did. Either way, this code is significantly easier to read than the previous version because it's simpler. I changed the name of "holder" to "number" to better reflect what is being compared.