Solution in mind, some work done
Took a while to understand the problem but it makes sense to me. I have an idea on how to solve the problem. I managed to add the extra instructions of 3 and 4 and tested them successfully. I managed to get a working solution of opcodes which are not 5 digits in length. The problem doesn't state it but it can be deduced that opcodes cannot go above 5 digits. When I say opcodes, I mean the instructions that actually get something done.  So if an instruction opcode is below 5 i can rest assurred that my future if statements will be able to figure out what to do. I am going to need to convert all instruction opcodes to strings to perform my operations. That will be interesting to try and debug. I stopped early to play terraria with poogie. Hoping my solution will result in an answer, we will see.